Fix bad conversion from seconds to ms
# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

+++ src/python/pants/engine/internals/native_engine.pyi
@@ -58,7 +58,7 @@ class PySnapshot:
 
 class PyExecutionRequest:
     def __init__(
-        self, poll: bool, poll_delay_in_ms: float | None, timeout_in_ms: float | None
+        self, poll: bool, poll_delay_in_ms: int | None, timeout_in_ms: int | None
     ) -> None: ...
 
 class PyExecutionStrategyOptions:
